# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atlas Backend API
  version: 0.1.0
  description: >
    OpenAPI documentation for the Atlas backend API. This spec covers the
    auth, users, voice, drafts, analytics, alerts, images, research, and
    trending routes exposed by the Express service.
servers:
  - url: /
    description: Current server
...
security:
  - bearerAuth: []
paths:
...
  /api/voice/profile:
    get:
      tags:
        - Voice
      summary: Get the authenticated user's voice profile
      responses:
        '200':
          description: Voice profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Voice
      summary: Update voice dimensions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoiceProfileRequest'
      responses:
        '200':
          description: Updated voice profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/voice/references:
    get:
      tags:
        - Voice
      summary: List active reference voices
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Reference voices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceVoicesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Voice
      summary: Create a reference voice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReferenceVoiceRequest'
      responses:
        '200':
          description: Reference voice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceVoiceEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/voice/blends:
    get:
      tags:
        - Voice
      summary: List saved blends
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Saved blends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlendsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Voice
      summary: Create a saved blend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlendRequest'
      responses:
        '200':
          description: Blend created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlendEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/voice/blends/{blendId}/voices/{voiceId}:
    patch:
      tags:
        - Voice
      summary: Update a voice entry inside a blend
      parameters:
        - $ref: '#/components/parameters/BlendIdParam'
        - $ref: '#/components/parameters/VoiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBlendVoiceRequest'
      responses:
        '200':
          description: Updated blend voice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlendVoiceEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Voice
      summary: Remove a voice entry from a blend
      parameters:
        - $ref: '#/components/parameters/BlendIdParam'
        - $ref: '#/components/parameters/VoiceIdParam'
      responses:
        '200':
          description: Blend voice deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/voice/calibrate:
    post:
      tags:
        - Voice
      summary: Calibrate the current voice profile from a Twitter handle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalibrateVoiceRequest'
      responses:
        '200':
          description: Voice profile calibrated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceCalibrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Max number of records to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    OffsetParam:
      name: offset
      in: query
      description: Number of records to skip
      schema:
        type: integer
        minimum: 0
        default: 0
...
    BlendIdParam:
      name: blendId
      in: path
      required: true
      schema:
        type: string
    VoiceIdParam:
      name: voiceId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadGateway:
      description: Upstream AI/provider request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - error
        - timestamp
      properties:
        ok:
          type: boolean
          enum:
            - false
        error:
          type: string
        details:
          nullable: true
          description: Additional validation or provider-specific detail payload.
        timestamp:
          type: string
          format: date-time
    SuccessEnvelope:
      type: object
      required:
        - ok
        - timestamp
      properties:
        ok:
          type: boolean
          enum:
            - true
        timestamp:
          type: string
          format: date-time
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - success
              properties:
                success:
                  type: boolean
...
    VoiceMaturity:
      type: string
      enum:
        - BEGINNER
        - INTERMEDIATE
        - ADVANCED
...
    VoiceProfile:
      type: object
      required:
        - id
        - userId
        - humor
        - formality
        - brevity
        - contrarianTone
        - maturity
        - tweetsAnalyzed
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        userId:
          type: string
        humor:
          type: integer
          minimum: 0
          maximum: 100
        formality:
          type: integer
          minimum: 0
          maximum: 100
        brevity:
          type: integer
          minimum: 0
          maximum: 100
        contrarianTone:
          type: integer
          minimum: 0
          maximum: 100
        maturity:
          $ref: '#/components/schemas/VoiceMaturity'
        tweetsAnalyzed:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ReferenceVoice:
      type: object
      required:
        - id
        - userId
        - name
        - isActive
        - createdAt
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        handle:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    BlendVoice:
      type: object
      required:
        - id
        - blendId
        - label
        - percentage
      properties:
        id:
          type: string
        blendId:
          type: string
        referenceVoiceId:
          type: string
          nullable: true
        label:
          type: string
        percentage:
          type: integer
        referenceVoice:
          allOf:
            - $ref: '#/components/schemas/ReferenceVoice'
          nullable: true
    SavedBlend:
      type: object
      required:
        - id
        - userId
        - name
        - voices
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        voices:
          type: array
          items:
            $ref: '#/components/schemas/BlendVoice'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TwitterUserSummary:
      type: object
      required:
        - username
        - name
      properties:
        username:
          type: string
        name:
          type: string
    VoiceCalibration:
      type: object
      required:
        - confidence
        - analysis
        - tweetsAnalyzed
        - twitterUser
      properties:
        confidence:
          type: number
          format: float
        analysis:
          type: string
        tweetsAnalyzed:
          type: integer
        twitterUser:
          $ref: '#/components/schemas/TwitterUserSummary'
...
    UpdateVoiceProfileRequest:
      type: object
      properties:
        humor:
          type: number
          minimum: 0
          maximum: 100
        formality:
          type: number
          minimum: 0
          maximum: 100
        brevity:
          type: number
          minimum: 0
          maximum: 100
        contrarianTone:
          type: number
          minimum: 0
          maximum: 100
    CreateReferenceVoiceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        handle:
          type: string
    WeightedBlendVoiceInput:
      type: object
      required:
        - referenceId
        - weight
      properties:
        referenceId:
          type: string
          minLength: 1
        weight:
          type: number
          minimum: 0
          maximum: 100
    LegacyBlendVoiceInput:
      type: object
      required:
        - label
        - percentage
      properties:
        label:
          type: string
          minLength: 1
        percentage:
          type: number
          minimum: 0
          maximum: 100
        referenceVoiceId:
          type: string
    CreateBlendRequest:
      type: object
      required:
        - name
        - voices
      properties:
        name:
          type: string
          minLength: 1
        voices:
          type: array
          minItems: 1
          items:
            oneOf:
              - $ref: '#/components/schemas/WeightedBlendVoiceInput'
              - $ref: '#/components/schemas/LegacyBlendVoiceInput'
    UpdateBlendVoiceRequest:
      type: object
      properties:
        label:
          type: string
        percentage:
          type: number
        referenceVoiceId:
          type: string
    CalibrateVoiceRequest:
      type: object
      required:
        - handle
      properties:
        handle:
          type: string
          minLength: 1
          maxLength: 50
...
    VoiceProfileResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - profile
              properties:
                profile:
                  $ref: '#/components/schemas/VoiceProfile'
    ReferenceVoicesResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - voices
              properties:
                voices:
                  type: array
                  items:
                    $ref: '#/components/schemas/ReferenceVoice'
    ReferenceVoiceEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - voice
              properties:
                voice:
                  $ref: '#/components/schemas/ReferenceVoice'
    BlendsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - blends
              properties:
                blends:
                  type: array
                  items:
                    $ref: '#/components/schemas/SavedBlend'
    BlendEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - blend
              properties:
                blend:
                  $ref: '#/components/schemas/SavedBlend'
    BlendVoiceEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - voice
              properties:
                voice:
                  $ref: '#/components/schemas/BlendVoice'
    VoiceCalibrationResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - profile
                - calibration
              properties:
                profile:
                  $ref: '#/components/schemas/VoiceProfile'
                calibration:
                  $ref: '#/components/schemas/VoiceCalibration'
...